01-anti-spider/requirements.txt
six
python-memcached
web.py
